requirements.txt
requests==2.27.1
virtualenv==20.14.1
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
python-decouple==3.6
requests==2.27.1
urllib3==1.26.9
